+++ packages/@aws-cdk/aws-sagemaker-alpha/test/endpoint.test.ts
@@ -1,5 +1,4 @@
 import * as cdk from 'aws-cdk-lib';
-import { Template, Match } from 'aws-cdk-lib/assertions';
 import * as sagemaker from '../lib';
 
 describe('When searching an Endpoint for a production variant', () => {
@@ -428,107 +427,5 @@ describe('When auto-scaling a production variant\'s instance count', () => {
 
     // THEN
     expect(when).toThrow(/Safety factor \(1\.1\) must be greater than 0.0 and less than or equal 1\.0/);
-  });
-});
-
-describe('When containerStartupHealthCheckTimeoutInSeconds is set', () => {
-  test('should be included in CloudFormation template when provided', () => {
-    // GIVEN
-    const stack = new cdk.Stack();
-    const model = sagemaker.Model.fromModelName(stack, 'Model', 'model');
-    // WHEN
-    new sagemaker.EndpointConfig(stack, 'EndpointConfig', {
-      instanceProductionVariants: [{
-        variantName: 'variant',
-        model,
-        containerStartupHealthCheckTimeout: cdk.Duration.minutes(5), // 300 seconds
-      }],
-    });
-    // THEN
-    Template.fromStack(stack).hasResourceProperties('AWS::SageMaker::EndpointConfig', {
-      ProductionVariants: [{
-        ModelName: 'model',
-        VariantName: 'variant',
-        ContainerStartupHealthCheckTimeoutInSeconds: 300,
-      }],
-    });
-  });
-  test('should not be included when not provided', () => {
-    // GIVEN
-    const stack = new cdk.Stack();
-    const model = sagemaker.Model.fromModelName(stack, 'Model', 'model');
-    // WHEN
-    new sagemaker.EndpointConfig(stack, 'EndpointConfig', {
-      instanceProductionVariants: [{
-        variantName: 'variant',
-        model,
-      }],
-    });
-    // THEN
-    Template.fromStack(stack).hasResourceProperties('AWS::SageMaker::EndpointConfig', {
-      ProductionVariants: [{
-        ModelName: 'model',
-        VariantName: 'variant',
-        ContainerStartupHealthCheckTimeoutInSeconds: Match.absent(),
-      }],
-    });
-  });
-
-  test('should throw error when timeout is less than 60 seconds', () => {
-    // GIVEN
-    const stack = new cdk.Stack();
-    const model = sagemaker.Model.fromModelName(stack, 'Model', 'model');
-    // WHEN & THEN
-    expect(() => {
-      new sagemaker.EndpointConfig(stack, 'EndpointConfig', {
-        instanceProductionVariants: [{
-          variantName: 'variant',
-          model,
-          containerStartupHealthCheckTimeout: cdk.Duration.seconds(30),
-        }],
-      });
-    }).toThrow('Configure \'containerStartupHealthCheckTimeout\' between 60 and 3600 seconds');
-  });
-
-  test('should throw error when timeout is greater than 3600 seconds', () => {
-    // GIVEN
-    const stack = new cdk.Stack();
-    const model = sagemaker.Model.fromModelName(stack, 'Model', 'model');
-    // WHEN & THEN
-    expect(() => {
-      new sagemaker.EndpointConfig(stack, 'EndpointConfig', {
-        instanceProductionVariants: [{
-          variantName: 'variant',
-          model,
-          containerStartupHealthCheckTimeout: cdk.Duration.hours(2), // 7200 seconds
-        }],
-      });
-    }).toThrow('Configure \'containerStartupHealthCheckTimeout\' between 60 and 3600 seconds');
-  });
-
-  test('should accept valid timeout values', () => {
-    // GIVEN
-    const stack = new cdk.Stack();
-    const model = sagemaker.Model.fromModelName(stack, 'Model', 'model');
-    // WHEN & THEN - should not throw
-    expect(() => {
-      new sagemaker.EndpointConfig(stack, 'EndpointConfig1', {
-        instanceProductionVariants: [{
-          variantName: 'variant',
-          model,
-          containerStartupHealthCheckTimeout: cdk.Duration.seconds(60), // minimum
-        }],
-      });
-    }).not.toThrow();
-
-    expect(() => {
-      new sagemaker.EndpointConfig(stack, 'EndpointConfig2', {
-        instanceProductionVariants: [{
-          variantName: 'variant',
-          model,
-          containerStartupHealthCheckTimeout: cdk.Duration.seconds(3600), // maximum
-        }],
-      });
-    }).not.toThrow();
   });
 });
